chore(root): update env and tsconfig
