+++ packages/cms/express-mini-apps/two-factor-auth/route.ts
@@ -8,7 +8,7 @@ import qrcode from 'qrcode'
 import { authenticator } from 'otplib'
 import { sign } from 'jsonwebtoken'
 
-// @ts-ignore
+// @ts-ignore twreporter errors is not typed
 import _errors from '@twreporter/errors'
 
 const cookieName2fa = appConfig.twoFactorAuth.cookieName
