+++ packages/frontend/.env.example
@@ -1,6 +1,6 @@
 # Graphql Endpoint
 INTERNAL_GQL_ENDPOINT=http://localhost:3000/api/graphql
-GQL_ENDPOINT=http://localhost:3000/api/graphql
+NEXT_PUBLIC_GQL_ENDPOINT=http://localhost:3000/api/graphql
 
-NEXT_PUBLIC_IS_PREVIEW_MODE=false
+# Release Env
 NEXT_PUBLIC_RELEASE_ENV=
