+++ packages/frontend/next.config.js
@@ -1,9 +1,8 @@
 /** @type {import('next').NextConfig} */
 
-const isPreviewMode = process.env.NEXT_PUBLIC_IS_PREVIEW_MODE === 'true'
-
 const nextConfig = {
-  basePath: isPreviewMode ? '/preview-server' : '',
+  basePath:
+    process.env.NEXT_PUBLIC_IS_PREVIEW_MODE === 'true' ? '/preview-server' : '',
   output: 'standalone',
   compiler: {
     styledComponents: true,
