+++ tsconfig.json
@@ -1,15 +1,7 @@
 {
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
     "isolatedModules": true,
-    "strict": true,
-    "module": "esnext",
-    "target": "esnext",
-    "jsx": "react-jsx",
-    "noEmit": true,
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "checkJs": true,
-    "allowJs": true
+    "checkJs": true
   }
 }
